VBA Script to process Stock data by year.
This script processes stock data by year and ticker summarizing the total volume, change in price and percent change over the year by Ticker.  Additionally, it tracks the Ticker with the greatest percent increase, greatest percent decrease and greatest total volume for the year.  Finally, it formats the data and the results.